requirements.txt
langchain
langgraph
langchain_tavily
langchain_openai
pydantic
pandas
beautifulsoup4
python-dotenv
accelerate